apps/ml/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
joblib==1.4.2
streamlit==1.44.1
python-dotenv==1.0.1
google-generativeai==0.8.4
fastapi==0.110.0
uvicorn==0.23.2
